+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYHAQQ42ZHAPKM9X92EKSXR9"
+        "01KYHFSEPYB7RHGM9S5YW820Y7"
     )
     r"""The unique identifier of the data source"""