+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYHAQQ3YB77E8FKQPD1P7GXX"
+        "01KYHFSEPSKYP213MWTWA6C08P"
     )
     r"""The unique identifier of the data source"""