+++ packages/orq-rc/src/orq_ai_sdk/models/modelcreatevertexop.py
@@ -12,7 +12,6 @@ from .modelparameterdocument import (
     ModelParameterDocumentTypedDict,
 )
 from .vertexconfiguration import VertexConfiguration, VertexConfigurationTypedDict
-from datetime import datetime
 from orq_ai_sdk.types import BaseModel, Nullable, UNSET_SENTINEL
 import pydantic
 from pydantic import model_serializer
@@ -35,7 +34,7 @@ class ModelCreateVertexResponseBodyTypedDict(TypedDict):
     r"""Vertex AI model created successfully"""
 
     configuration: ModelConfigurationResponseTypedDict
-    created: datetime
+    created: str
     description: Nullable[str]
     display_name: str
     docs_url: Nullable[str]
@@ -58,7 +57,7 @@ class ModelCreateVertexResponseBodyTypedDict(TypedDict):
     pricing_url: Nullable[str]
     provider: str
     ref_id: str
-    updated: datetime
+    updated: str
     model_developer: NotRequired[str]
     model_family: NotRequired[str]
     sharing: NotRequired[ConfigTypedDict]
@@ -69,7 +68,7 @@ class ModelCreateVertexResponseBody(BaseModel):
 
     configuration: ModelConfigurationResponse
 
-    created: datetime
+    created: str
 
     description: Nullable[str]
 
@@ -115,7 +114,7 @@ class ModelCreateVertexResponseBody(BaseModel):
 
     ref_id: Annotated[str, pydantic.Field(alias="refId")]
 
-    updated: datetime
+    updated: str
 
     model_developer: Optional[str] = None