+++ packages/orq-rc/src/orq_ai_sdk/models/modeldocument.py
@@ -11,7 +11,6 @@ from .modelparameterdocument import (
     ModelParameterDocument,
     ModelParameterDocumentTypedDict,
 )
-from datetime import datetime
 from orq_ai_sdk.types import BaseModel, Nullable, UNSET_SENTINEL
 import pydantic
 from pydantic import model_serializer
@@ -21,7 +20,7 @@ from typing_extensions import Annotated, NotRequired, TypedDict
 
 class ModelDocumentTypedDict(TypedDict):
     configuration: ModelConfigurationResponseTypedDict
-    created: datetime
+    created: str
     description: Nullable[str]
     display_name: str
     docs_url: Nullable[str]
@@ -44,7 +43,7 @@ class ModelDocumentTypedDict(TypedDict):
     pricing_url: Nullable[str]
     provider: str
     ref_id: str
-    updated: datetime
+    updated: str
     model_developer: NotRequired[str]
     model_family: NotRequired[str]
     sharing: NotRequired[ConfigTypedDict]
@@ -53,7 +52,7 @@ class ModelDocumentTypedDict(TypedDict):
 class ModelDocument(BaseModel):
     configuration: ModelConfigurationResponse
 
-    created: datetime
+    created: str
 
     description: Nullable[str]
 
@@ -99,7 +98,7 @@ class ModelDocument(BaseModel):
 
     ref_id: Annotated[str, pydantic.Field(alias="refId")]
 
-    updated: datetime
+    updated: str
 
     model_developer: Optional[str] = None