+++ packages/orq-rc/src/orq_ai_sdk/models/modelupdateawsbedrockop.py
@@ -11,7 +11,6 @@ from .modelparameterdocument import (
     ModelParameterDocument,
     ModelParameterDocumentTypedDict,
 )
-from datetime import datetime
 from orq_ai_sdk.types import BaseModel, Nullable, UNSET_SENTINEL
 from orq_ai_sdk.utils import FieldMetadata, PathParamMetadata, RequestMetadata
 import pydantic
@@ -154,7 +153,7 @@ class ModelUpdateAwsBedrockResponseBodyTypedDict(TypedDict):
     r"""AWS Bedrock model updated successfully"""
 
     configuration: ModelConfigurationResponseTypedDict
-    created: datetime
+    created: str
     description: Nullable[str]
     display_name: str
     docs_url: Nullable[str]
@@ -177,7 +176,7 @@ class ModelUpdateAwsBedrockResponseBodyTypedDict(TypedDict):
     pricing_url: Nullable[str]
     provider: str
     ref_id: str
-    updated: datetime
+    updated: str
     model_developer: NotRequired[str]
     model_family: NotRequired[str]
     sharing: NotRequired[ConfigTypedDict]
@@ -188,7 +187,7 @@ class ModelUpdateAwsBedrockResponseBody(BaseModel):
 
     configuration: ModelConfigurationResponse
 
-    created: datetime
+    created: str
 
     description: Nullable[str]
 
@@ -234,7 +233,7 @@ class ModelUpdateAwsBedrockResponseBody(BaseModel):
 
     ref_id: Annotated[str, pydantic.Field(alias="refId")]
 
-    updated: datetime
+    updated: str
 
     model_developer: Optional[str] = None