+++ packages/orq-rc/src/orq_ai_sdk/models/modelupdateopenailikeop.py
@@ -11,7 +11,6 @@ from .modelparameterdocument import (
     ModelParameterDocument,
     ModelParameterDocumentTypedDict,
 )
-from datetime import datetime
 from orq_ai_sdk.types import BaseModel, Nullable, UNSET_SENTINEL
 from orq_ai_sdk.utils import FieldMetadata, PathParamMetadata, RequestMetadata
 import pydantic
@@ -127,7 +126,7 @@ class ModelUpdateOpenAILikeResponseBodyTypedDict(TypedDict):
     r"""OpenAI-like model updated successfully"""
 
     configuration: ModelConfigurationResponseTypedDict
-    created: datetime
+    created: str
     description: Nullable[str]
     display_name: str
     docs_url: Nullable[str]
@@ -150,7 +149,7 @@ class ModelUpdateOpenAILikeResponseBodyTypedDict(TypedDict):
     pricing_url: Nullable[str]
     provider: str
     ref_id: str
-    updated: datetime
+    updated: str
     model_developer: NotRequired[str]
     model_family: NotRequired[str]
     sharing: NotRequired[ConfigTypedDict]
@@ -161,7 +160,7 @@ class ModelUpdateOpenAILikeResponseBody(BaseModel):
 
     configuration: ModelConfigurationResponse
 
-    created: datetime
+    created: str
 
     description: Nullable[str]
 
@@ -207,7 +206,7 @@ class ModelUpdateOpenAILikeResponseBody(BaseModel):
 
     ref_id: Annotated[str, pydantic.Field(alias="refId")]
 
-    updated: datetime
+    updated: str
 
     model_developer: Optional[str] = None