+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYHAQQ41SACYSH7RY7SSF516"
+        "01KYHFSEPWERTGKZ4SYMZQQ639"
     )
     r"""The unique identifier of the data source"""