+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYHAQQ46AR1BX4641ZD5JW1Q"
+        "01KYHFSEQ3C33SA1ZT94MSY4F7"
     )
     r"""The unique identifier of the data source"""