+++ packages/orq-rc/RELEASES.md
@@ -9749,3 +9749,13 @@ Based on:
 - [python v4.13.0-rc.6] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.6 - packages/orq-rc
+
+## 2026-07-27 09:53:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.7] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.7 - packages/orq-rc